Code社区

+++ src/main/java/com/bbgu/zmz/community/service/TopicService.java
@@ -1,16 +1,15 @@
 package com.bbgu.zmz.community.service;
 
-import com.bbgu.zmz.community.dto.ReplyDTO;
 import com.bbgu.zmz.community.dto.Result;
-import com.bbgu.zmz.community.dto.TopicInfoDTO;
 import com.bbgu.zmz.community.enums.MsgEnum;
 import com.bbgu.zmz.community.mapper.*;
+import com.bbgu.zmz.community.mapper.TopicinfoExtMapper;
 import com.bbgu.zmz.community.model.*;
+import com.bbgu.zmz.community.model.TopicinfoExt;
 import com.bbgu.zmz.community.util.StringDate;
-import org.apache.ibatis.session.RowBounds;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import java.util.ArrayList;
+
 import java.util.Date;
 import java.util.List;
 
@@ -35,24 +34,21 @@ public class TopicService {
     private MessageMapper messageMapper;
     @Autowired
     private CommentagreeMapper commentagreeMapper;
+    @Autowired
+    private CommentExtMapper commentExtMapper;
 
 
     /*
-    获取置顶帖子信息，查询4条
+    获取置顶帖子信息 || 获取最新综合帖子信息
      */
-    public List<TopicInfoDTO> topicTop(Integer istop,Integer num){
-        List<TopicInfoDTO> topicInfoDTOList = new ArrayList<TopicInfoDTO>();
-        //获取帖子信息
-        TopicinfoExample topicinfoExample = new TopicinfoExample();
-        topicinfoExample.createCriteria().andIsTopEqualTo(istop);
-        topicinfoExample.setOrderByClause("topic_create desc");
-        List<Topicinfo> topicinfos = topicinfoMapper.selectByExampleWithRowbounds(topicinfoExample,new RowBounds(0,num));
-        for (Topicinfo topicinfo:topicinfos) {
-            TopicInfoDTO topicInfoDTO  =  getDetails(topicinfo);
-            topicInfoDTOList.add(topicInfoDTO);
+    public List<TopicinfoExt> topicTop(Integer istop,Integer offset,Integer size){
+        List<TopicinfoExt> topicinfoExtList =  topicinfoExtMapper.getTopTopic(istop,offset,size);
+        for(TopicinfoExt topicinfoExt:topicinfoExtList){
+            topicinfoExt.setTime(StringDate.getStringDate(new Date(topicinfoExt.getTopicCreate())));
         }
-        return topicInfoDTOList;
+        return topicinfoExtList;
     }
+
     /*
     查询一级分类
      */
@@ -81,42 +77,15 @@ public class TopicService {
     }
 
     /*
-    查看详情
+    查看帖子详情
      */
-    public TopicInfoDTO showDetail(Long id){
-        //获得帖子信息
-        Topicinfo topicinfo = topicinfoMapper.selectByPrimaryKey(id);
-        TopicInfoDTO topicInfoDTO = getDetails(topicinfo);
-        return topicInfoDTO;
+    public TopicinfoExt showDetail(Long id){
+        TopicinfoExt topicinfoExt = topicinfoExtMapper.getTopicDetails(id);
+        topicinfoExt.setTime(StringDate.getStringDate(new Date(topicinfoExt.getTopicCreate())));
+        return topicinfoExt;
+
     }
-    /*
-    获取帖子相关信息
-     */
-    public TopicInfoDTO getDetails(Topicinfo topicinfo){
-        TopicInfoDTO topicInfoDTO = new TopicInfoDTO();
-        topicInfoDTO.setTopicinfo(topicinfo);
-        //获取时间
-        String time = StringDate.getStringDate(new Date(topicinfo.getTopicCreate()));
-        topicInfoDTO.setTime(time);
-        //获取用户信息
-        UserExample userExample = new UserExample();
-        userExample.createCriteria().andAccountIdEqualTo(topicinfo.getUserId());
-        List<User> users = userMapper.selectByExample(userExample);
-        User user = users.get(0);
-        topicInfoDTO.setUser(user);
-        //获取帖子分类
-        Category category = categoryMapper.selectByPrimaryKey(topicinfo.getCategoryId());
-        topicInfoDTO.setCategory(category);
-        //获取帖子二级分类
-        Kind kind = kindMapper.selectByPrimaryKey(topicinfo.getKindId());
-        topicInfoDTO.setKind(kind);
-        //获得评论总数
-        CommentExample commentExample = new CommentExample();
-        commentExample.createCriteria().andTopicIdEqualTo(topicinfo.getId());
-        Long comment_count = commentMapper.countByExample(commentExample);
-        topicInfoDTO.setComment_count(comment_count);
-        return topicInfoDTO;
-    }
+
     /*
     更新浏览数
      */
@@ -134,87 +103,27 @@ public class TopicService {
     /*
     查询评论
      */
-    public List<ReplyDTO> findComment(Long id,Integer page,Integer size,User userinfo){
-        List<ReplyDTO> replyDTOList = new ArrayList<ReplyDTO>();
-        CommentExample commentExample = new CommentExample();
-        commentExample.createCriteria().andTopicIdEqualTo(id);
-        Long count = commentMapper.countByExample(commentExample);
-        Long totalPage;
-        if(count % size ==0){
-            totalPage = count / size;
-        }else{
-            totalPage = count / size + 1;
-        }
-        if(page < 1){
-            page=1;
-        }
-        if(page > totalPage){
-            page = Integer.parseInt(String.valueOf(totalPage));
-        }
-        Integer offset = size * (page - 1);
-       // commentExample.setOrderByClause("comment_create desc");
-        List<Comment> commentList = commentMapper.selectByExampleWithRowbounds(commentExample,new RowBounds(offset,size));
+    public List<CommentExt> findComment(Long id,Integer page,Integer size,User userinfo){
+        Integer offset = (page -1)*size;
+       List<CommentExt> commentExtList = commentExtMapper.getCommentByTid(id,offset,size);
 
-       for(Comment comment:commentList){
-           ReplyDTO replyDTO = new ReplyDTO();
-           String time = StringDate.getStringDate(new Date(comment.getCommentCreate()));
-           replyDTO.setComment(comment);
-           UserExample userExample = new UserExample();
-           userExample.createCriteria().andAccountIdEqualTo(comment.getUserId());
-           List<User> users = userMapper.selectByExample(userExample);
-           User user = users.get(0);
-           //查询是否点赞
-           if(userinfo != null){
-               CommentagreeExample commentagreeExample = new CommentagreeExample();
-               commentagreeExample.createCriteria().andUseridEqualTo(userinfo.getAccountId()).andCommentIdEqualTo(comment.getId());
-               List<Commentagree> commentagreeList = commentagreeMapper.selectByExample(commentagreeExample);
-               if(commentagreeList.size()>0){
-                   replyDTO.setZan(true);
-               }else{
-                   replyDTO.setZan(false);
-               }
-           }
-           replyDTO.setCount(count);
-           replyDTO.setTime(time);
-           replyDTO.setUser(user);
-           replyDTOList.add(replyDTO);
-       }
-        return  replyDTOList;
-    }
-    /*
-    查询采纳评论
-     */
-    public ReplyDTO findAcceptComment(Long id,User userinfo){
-        CommentExample commentExample = new CommentExample();
-        commentExample.createCriteria().andTopicIdEqualTo(id).andIsAcceptEqualTo(1);
-        List<Comment> comments =commentMapper.selectByExample(commentExample);
-        if(comments.size() != 0) {
-            ReplyDTO replyDTO = new ReplyDTO();
-            Comment comment = comments.get(0);
-            if(userinfo != null){
-                CommentagreeExample commentagreeExample = new CommentagreeExample();
-                commentagreeExample.createCriteria().andUseridEqualTo(userinfo.getAccountId()).andCommentIdEqualTo(comment.getId());
-                List<Commentagree> commentagreeList = commentagreeMapper.selectByExample(commentagreeExample);
-                if(commentagreeList.size()>0){
-                    replyDTO.setZan(true);
-                }else{
-                    replyDTO.setZan(false);
-                }
+        for(CommentExt commentExt:commentExtList){
+            //查询是否点赞
+        if(userinfo != null){
+            CommentagreeExample commentagreeExample = new CommentagreeExample();
+            commentagreeExample.createCriteria().andUseridEqualTo(userinfo.getAccountId()).andCommentIdEqualTo(commentExt.getId());
+            List<Commentagree> commentagreeList = commentagreeMapper.selectByExample(commentagreeExample);
+            if(commentagreeList.size()>0){
+                commentExt.setZan(true);
+            }else{
+                commentExt.setZan(false);
             }
-            replyDTO.setComment(comment);
-            UserExample userExample = new UserExample();
-            userExample.createCriteria().andAccountIdEqualTo(comment.getUserId());
-            List<User> users = userMapper.selectByExample(userExample);
-            User user = users.get(0);
-            replyDTO.setCount(99L);
-            replyDTO.setUser(user);
-            return replyDTO;
         }
-        return null;
+        //转换时间
+        commentExt.setTime(StringDate.getStringDate(new Date(commentExt.getCommentCreate())));
     }
-
-
-
+        return commentExtList;
+    }
 
     /*
     审核帖子
@@ -267,45 +176,44 @@ public class TopicService {
     点赞
      */
     public Result jiedaZan(User user, Long id, Boolean ok){
-       // RegRespObj regRespObj = new RegRespObj();
         if(user != null)
         {
             if(ok == false)//点赞
             {
-                Comment comment = commentMapper.selectByPrimaryKey(id);
-                comment.setAgreeNum(comment.getAgreeNum() + 1);
-                commentMapper.updateByPrimaryKeySelective(comment);
-                Commentagree commentagree = new Commentagree();
-                commentagree.setCommentId(id);
-                commentagree.setUserid(user.getAccountId());
-                commentagreeMapper.insertSelective(commentagree);
-/*                regRespObj.setStatus(0);
-                regRespObj.setMsg("点赞成功！");*/
-                return new Result().ok(MsgEnum.ZAN_SUCCESS);
-            }else{  //取消点赞
-                Comment comment = commentMapper.selectByPrimaryKey(id);
-                comment.setAgreeNum(comment.getAgreeNum() - 1);
-                commentMapper.updateByPrimaryKeySelective(comment);
                 CommentagreeExample commentagreeExample = new CommentagreeExample();
                 commentagreeExample.createCriteria().andCommentIdEqualTo(id).andUseridEqualTo(user.getAccountId());
-                commentagreeMapper.deleteByExample(commentagreeExample);
-/*                regRespObj.setStatus(0);
-                regRespObj.setMsg("取消成功！");*/
-                return new Result().ok(MsgEnum.ZAN_CANCEL);
+                List<Commentagree> commentagreeList  =commentagreeMapper.selectByExample(commentagreeExample);
+                if(commentagreeList.size() == 0){
+                    commentExtMapper.updateAgreeNumAdd(id);
+                    Commentagree commentagree = new Commentagree();
+                    commentagree.setCommentId(id);
+                    commentagree.setUserid(user.getAccountId());
+                    commentagreeMapper.insertSelective(commentagree);
+                    return new Result().ok(MsgEnum.ZAN_SUCCESS);
+                }else{
+                    return new Result().error(MsgEnum.ZAN_FAILE);
+                }
+            }else{  //取消点赞
+                CommentagreeExample commentagreeExample = new CommentagreeExample();
+                commentagreeExample.createCriteria().andCommentIdEqualTo(id).andUseridEqualTo(user.getAccountId());
+                List<Commentagree> commentagreeList  =commentagreeMapper.selectByExample(commentagreeExample);
+                if(commentagreeList.size() == 0){
+                    return new Result().error(MsgEnum.ZAN_FAILE);
+                }else{
+                    commentExtMapper.updateAgreeNumSub(id);
+                    commentagreeMapper.deleteByExample(commentagreeExample);
+                    return new Result().ok(MsgEnum.ZAN_CANCEL);
+                }
             }
         }else{
-/*            regRespObj.setStatus(1);
-            regRespObj.setMsg("请登陆后再操作！");*/
             return new Result().error(MsgEnum.NOTLOGIN);
         }
-   // return regRespObj;
     }
     /*
     采纳评论
      */
 
-    public Result acceptComment(Long id,User user){
-        //RegRespObj regRespObj = new RegRespObj();
+    public Result acceptComment(Long id){
         Comment comment = new Comment();
         comment.setId(id);
         comment.setIsAccept(1);
@@ -315,8 +223,6 @@ public class TopicService {
         topicinfo.setId(comment.getTopicId());
         topicinfo.setIsEnd(1);
         topicinfoMapper.updateByPrimaryKeySelective(topicinfo);
-      /*  regRespObj.setStatus(0);
-        return regRespObj;*/
         return new Result().ok(MsgEnum.OK);
     }
 
@@ -337,10 +243,6 @@ public class TopicService {
         comment.setContent(content);
         comment.setCommentModified(System.currentTimeMillis());
         commentMapper.updateByPrimaryKeySelective(comment);
-/*        RegRespObj regRespObj = new RegRespObj();
-        regRespObj.setStatus(0);
-        regRespObj.setMsg("评论已更新！");
-        return regRespObj;*/
         return new Result().ok(MsgEnum.UPDATE_COMMENT);
     }
 
@@ -357,11 +259,7 @@ public class TopicService {
         CommentagreeExample commentagreeExample = new CommentagreeExample();
         commentagreeExample.createCriteria().andCommentIdEqualTo(id);
         commentagreeMapper.deleteByExample(commentagreeExample);
-/*        RegRespObj regRespObj = new RegRespObj();
-        regRespObj.setStatus(0);
-        regRespObj.setMsg("删除成功！");
-        return regRespObj;*/
-        return new Result().ok(MsgEnum.DELETE_COMMENT);
+        return new Result().ok(MsgEnum.DELETE_SUCCESS);
     }
     /*
     修改帖子信息
@@ -523,7 +421,7 @@ public class TopicService {
     /*
     搜索
      */
-    public List<TopicinfoExt> searchTopic(String q,Integer offset, Integer size){
+    public List<TopicinfoExt> searchTopic(String q, Integer offset, Integer size){
         return topicinfoExtMapper.searchTopic(q,offset,size);
     }
